ref: move bitbucket_server testutils to fixtures
