+++ tests/sentry/integrations/bitbucket_server/test_integration.py
@@ -1,11 +1,10 @@
 import responses
 from requests.exceptions import ReadTimeout
 
+from fixtures.bitbucket_server import EXAMPLE_PRIVATE_KEY
 from sentry.integrations.bitbucket_server import BitbucketServerIntegrationProvider
 from sentry.models import Identity, IdentityProvider, Integration, OrganizationIntegration
 from sentry.testutils import IntegrationTestCase
-
-from .testutils import EXAMPLE_PRIVATE_KEY
 
 
 class BitbucketServerIntegrationTest(IntegrationTestCase):
